Additional move class to fit the two benchmarks
