+++ src/test/java/br/unb/cic/mop/bench01/KeyGeneratorTest.java
@@ -1,4 +1,4 @@
-package br.unb.cic.mop.jca;
+package br.unb.cic.mop.bench01;
 
 import br.unb.cic.mop.eh.ErrorCollector;
