+++ src/test/java/br/unb/cic/mop/bench01/KeyManagerFactoryTest.java
@@ -1,4 +1,4 @@
-package br.unb.cic.mop.jca;
+package br.unb.cic.mop.bench01;
 
 import org.junit.Before;
 import org.junit.Ignore;
